% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzle-conversion.R
\name{convert_puzzle_string_to_puzzle}
\alias{convert_puzzle_string_to_puzzle}
\alias{convert_puzzle_to_puzzle_string}
\title{Convert between nonogram puzzle strings and puzzles}
\usage{
convert_puzzle_string_to_puzzle(puzzle_string)

convert_puzzle_to_puzzle_string(puzzle)
}
\arguments{
\item{puzzle_string}{compact string representation of a puzzle}

\item{puzzle}{R list representation of a puzzle}
}
\description{
A 'puzzle string' is a compact text representation of a nonogram puzzle, whereas a
'puzzle' is an R list object representing the puzzle.
}
\details{
A \code{puzzle} is a list containing:
\itemize{
  \item{a list of the row clues}{}
  \item{a list of the column clues}{}
}

A \code{clue} is a list of integers which is the run-length encoding of just
the filled spaces in the final puzzle.

e.g. \code{
list(row_clues = list(2L, 2:1, c(1L, 1L), 3L, c(1L, 1L), c(1L, 1L), 2L, c(1L, 1L), 1:2, 2L),
     col_clues = list(2:1, c(2L, 1L, 3L), 7L, c(1L, 3L), 2:1))
}

A \code{puzzle string} is a just a compact string representation of the puzzle.
It consists of:
\itemize{
  \item{the numbers for each clue separated by a comma}{}
  \item{each clue separated by a colon}{}
  \item{the row and column clues (in that order) separated by a dash}{}
}

e.g. \code{"2:2,1:1,1:3:1,1:1,1:2:1,1:1,2:2-2,1:2,1,3:7:1,3:2,1"}
}
