% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzle-creation.R
\name{create_clue_from_pattern}
\alias{create_clue_from_pattern}
\title{Given a vector of 0/1 values, create the corresponding nonogram clue}
\usage{
create_clue_from_pattern(pattern)
}
\arguments{
\item{pattern}{vector of 0/1 values}
}
\value{
corresponding nonogram clue
}
\description{
Given a vector of 0/1 values, create the corresponding nonogram clue
}
