% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_clue_plot}
\alias{create_clue_plot}
\title{Create plot of all possibilities for a given clue}
\usage{
create_clue_plot(clue, total_length, row_spacing = 0.3, ...)
}
\arguments{
\item{clue}{nongram clue}

\item{total_length}{the total length the clue spans}

\item{row_spacing}{fractional extra spacing between rows}

\item{...}{arguments passed to \code{create_matrix_plot}}
}
\value{
ggplot2 object
}
\description{
Create plot of all possibilities for a given clue
}
