% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-pattern-set-from-clue.R
\name{create_pattern_set}
\alias{create_pattern_set}
\title{Create a matrix of all possible patterns for the specified clue}
\usage{
create_pattern_set(clue, total_length)
}
\arguments{
\item{clue}{run-length encoding of just the ones}

\item{total_length}{length of the pattern}
}
\value{
matrix with one pattern per row
}
\description{
Create a matrix of all possible patterns for the specified clue
}
