% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzle-creation.R
\name{create_puzzle_from_matrix}
\alias{create_puzzle_from_matrix}
\title{Given a matrix, create a nonogram puzzle}
\usage{
create_puzzle_from_matrix(mat)
}
\arguments{
\item{mat}{integer matrix with just 0/1 values}
}
\value{
puzzle in R object format i.e. a list
        of \code{row_clues} and \code{col_clues}
}
\description{
Given a matrix, create a nonogram puzzle
}
