% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_puzzle_plot}
\alias{create_puzzle_plot}
\title{Plot a nonogram puzzle}
\usage{
create_puzzle_plot(puzzle, solution_matrix = NULL, show_clues = TRUE,
  title = NULL, base_size = 11, col_spacing = 0, row_spacing = 0)
}
\arguments{
\item{puzzle}{puzzle object}

\item{solution_matrix}{solution matrix. optional}

\item{show_clues}{whether or not to show the clues along the axes. default: TRUE}

\item{title}{plot title. default: NULL}

\item{base_size}{base size for ggplot theme. default 11}

\item{col_spacing}{fractional extra spacing between columns}

\item{row_spacing}{fractional extra spacing between rows}
}
\value{
ggplot2 object
}
\description{
Plot a nonogram puzzle
}
