% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.nonogram.R
\name{dim.nonogram}
\alias{dim.nonogram}
\title{dim.nonogram}
\usage{
\method{dim}{nonogram}(nonogram)
}
\arguments{
\item{nonogram}{An object of class nonogram.}
}
\value{
Vector of length 2.
}
\description{
Returns the dimensions of a nonogram object.
}
\examples{
f <- examples[[31]]
dim(f)
}
