% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examples}
\alias{examples}
\title{Example nonograms}
\format{
A list, each element of which is a nonogram object.
}
\usage{
examples
}
\description{
A list of more than 100 nonograms of various dimensions.
}
\keyword{datasets}
