% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{filter_pattern_sets}
\alias{filter_pattern_sets}
\title{Given pattern sets for one axis, only keep those that match an
orthogonal pattern set at the given index}
\usage{
filter_pattern_sets(axis_pattern_sets, orthog_pattern_set, idx)
}
\arguments{
\item{axis_pattern_sets}{pattern sets for an axis}

\item{orthog_pattern_set}{pattern sets for an axis so far}

\item{idx}{which element to look at}
}
\value{
filtered version of axis_pattern_sets
}
\description{
Given pattern sets for one axis, only keep those that match an
orthogonal pattern set at the given index
}
