% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{filter_with_orthogonal_pattern_sets}
\alias{filter_with_orthogonal_pattern_sets}
\title{Filter pattern sets for one axis with the pattern sets for the other axis}
\usage{
filter_with_orthogonal_pattern_sets(axis_pattern_sets, orthog_pattern_sets)
}
\arguments{
\item{axis_pattern_sets}{first set}

\item{orthog_pattern_sets}{second set}
}
\value{
axis_pattern_sets filtered by the constraints of the orthogonal pattern sets
}
\description{
Filter pattern sets for one axis with the pattern sets for the other axis
}
