% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-integers-with-sum.R
\name{find_integers_with_sum}
\alias{find_integers_with_sum}
\title{Find all permutations of all vectors of positive integers of a given length which sum to a given value}
\usage{
find_integers_with_sum(target_sum, target_length, current_vec = c())
}
\arguments{
\item{target_sum}{target sum}

\item{target_length}{target length}

\item{current_vec}{current working set for the back-tracking solution process}
}
\value{
list of integer vectors
}
\description{
e.g. \code{find_integers_with_sum(4, 2)} gives `c(1, 3)`, `c(2, 2)`, `c(3, 1)`
}
