% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-integers-with-sum.R
\name{find_integers_with_sum_for_nonograms}
\alias{find_integers_with_sum_for_nonograms}
\title{Find all permutations of all vectors of positive numbers of a given length
which sum to a given value, with zeros allowed as first and last elements}
\usage{
find_integers_with_sum_for_nonograms(target_sum, target_length)
}
\arguments{
\item{target_sum}{target sum}

\item{target_length}{target length}
}
\value{
matrix
}
\description{
This is used to find all possible run-length encodings of the patterns
of zeros for a nonogram clue.
}
