% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_full_perm_set.R
\name{make_full_perm_set}
\alias{make_full_perm_set}
\title{make_full_perm_set}
\usage{
make_full_perm_set(length)
}
\arguments{
\item{length}{The length of each permutation.}
}
\value{
Returns a matrix with n columns, where each row corresponds to a
different permutation of 0s and 1s.
}
\description{
Given a length n, \code{make_full_perm_set} finds all permutations of
length n whose elements can be 0 or 1.
}
\examples{
# All possible permutations of length 3
make_full_perm_set(3)
}
