% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonogram.R
\name{nonogram}
\alias{nonogram}
\title{nonogram}
\usage{
nonogram(rows, columns)
}
\arguments{
\item{rows}{A list whose elements are numeric vectors indicating the pattern
within each row (e.g. \code{c(2, 1)} indicates a row with a run of length 2
followed by a run of length 1). The first element in the list corresponds
to the top row of the nonogram.}

\item{columns}{A list whose elements are numeric vectors indicating the pattern
within each column (e.g. \code{c(2, 1)} indicates a column with a run of length 2
followed by a run of length 1). The first element in the list corresponds
to the leftmost column of the nonogram.}
}
\value{
A nonogram object containing the found solution.
}
\description{
Constructor function to create an instance of the S3 class nonogram
}
\examples{
nonogram(
    rows    = list(c(1), c(1), c(2, 1), c(2, 1), c(3, 2), c(4, 3), c(5, 4),  c(1),  c(8),  c(6)),
    columns = list(c(1), c(2, 1),  c(3, 2),  c(5, 2),  c(7, 2),  c(3),  c(5, 2),  c(3, 2),  c(2, 1),  c(1))
   )
}
