% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_solver.R
\name{permutation_solver}
\alias{permutation_solver}
\title{permutation_solver}
\usage{
permutation_solver(pattern, permutation_patterns)
}
\arguments{
\item{pattern}{A numeric vector where each element corresponds to the length of a run.}

\item{permutation_patterns}{A matrix of permutations as generated by \code{make_full_perm_set()}}
}
\value{

}
\description{
Given a row or column pattern and the set of all possible
permutations of the row/column's length, \code{permutation_solver} identifies
the set of permutations that satisfy the pattern
}
\examples{
full <- make_full_perm_set(10)
permutation_solver(c(2, 4, 1), full)
}
