% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nonogram.R
\name{print.nonogram}
\alias{print.nonogram}
\title{print.nonogram}
\usage{
\method{print}{nonogram}(nonogram)
}
\arguments{
\item{nonogram}{An object of class \code{nonogram}.}
}
\description{
Simply prints the dimensions of the nonogram and whether it has
been solved. If the nonogram has been solved, the solution matrix is also
printed.
}
\examples{
f <- nonogram(
    rows = list(c(1), c(1), c(2, 1), c(2, 1), c(3, 2), c(4, 3), c(5, 4),  c(1),  c(8),  c(6)),
    columns = list(c(1), c(2, 1),  c(3, 2),  c(5, 2),  c(7, 2),  c(3),  c(5, 2),  c(3, 2),  c(2, 1),  c(1))
   )
   
print(f)

f <- solve(f)

print(f)
}
