% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzle-strings.R
\docType{data}
\name{puzzle_string_examples}
\alias{puzzle_string_examples}
\title{Test puzzle strings used during development - R, duck, swan, squirrel and goose}
\format{An object of class \code{list} of length 5.}
\usage{
puzzle_string_examples
}
\description{
A \code{puzzle string} is a just a compact string representation of the puzzle.
It consists of:
\itemize{
  \item{the numbers for each clue separated by a comma}{}
  \item{each clue separated by a colon}{}
  \item{the row and column clues (in that order) separated by a dash}{}
}
}
\details{
e.g. \code{"2:2,1:1,1:3:1,1:1,1:2:1,1:1,2:2-2,1:2,1,3:7:1,3:2,1"}
}
\keyword{datasets}
