% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzle-strings.R
\docType{data}
\name{puzzle_string_library}
\alias{puzzle_string_library}
\title{A collection of a few thousand nonogram puzzle strings}
\format{An object of class \code{character} of length 4239.}
\usage{
puzzle_string_library
}
\description{
A \code{puzzle string} is a just a compact string representation of the puzzle.
It consists of:
\itemize{
  \item{the numbers for each clue separated by a comma}{}
  \item{each clue separated by a colon}{}
  \item{the row and column clues (in that order) separated by a dash}{}
}
}
\details{
e.g. \code{"2:2,1:1,1:3:1,1:1,1:2:1,1:1,2:2-2,1:2,1,3:7:1,3:2,1"}
}
\keyword{datasets}
