% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve_puzzle}
\alias{solve_puzzle}
\title{Solve a puzzle by searching recursive with back-tracking and pruning.}
\usage{
solve_puzzle(puzzle, verbose = FALSE)
}
\arguments{
\item{puzzle}{nonogram puzzle or puzzle string}

\item{verbose}{print out timing information. default: FALSE}
}
\value{
solution as a matrix
}
\description{
Solve a puzzle by searching recursive with back-tracking and pruning.
}
