% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve_puzzle_core}
\alias{solve_puzzle_core}
\title{Solve a puzzle by searching recursive with back-tracking and pruning.}
\usage{
solve_puzzle_core(row_solution_patterns, col_pattern_sets, row_pattern_sets)
}
\arguments{
\item{row_solution_patterns}{solution so far of row patterns}

\item{col_pattern_sets}{column pattern sets}

\item{row_pattern_sets}{row pattern sets}
}
\value{
solution
}
\description{
Solve a puzzle by searching recursive with back-tracking and pruning.
}
