% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.nonogram.R
\name{as.data.frame.nonogram}
\alias{as.data.frame.nonogram}
\title{as.data.frame.nonogram}
\usage{
\method{as.data.frame}{nonogram}(nonogram, with_solution = TRUE)
}
\arguments{
\item{nonogram}{}

\item{with_solution}{}
}
\value{

}
\description{
Extracts the solution of a solved nonogram into a tidy data.frame.
}
