% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_solver.R
\name{permutation_solver}
\alias{permutation_solver}
\title{permutation_solver}
\usage{
permutation_solver(pattern, full_perms)
}
\arguments{
\item{pattern}{}

\item{permutation_patterns}{}
}
\value{

}
\description{
Given a row or column pattern, a set of candidate patterns,
and the set of all possible permutations of the row/column's length,
\code{permutation_solver} identifies the set of permutations that satisfy the
pattern TODO: edit this so it takes just the true pattern and the permutations, and include the
rle step inside it!
}
