% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_solver.R
\name{permutation_solver}
\alias{permutation_solver}
\title{permutation_solver}
\usage{
permutation_solver(pattern, permutation_patterns, full_perms)
}
\arguments{
\item{pattern}{}

\item{permutation_patterns}{}
}
\value{

}
\description{
I'm pretty sure this is important but I don't remember what it does or how it works...
}
