% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_solver.R
\name{permutation_solver}
\alias{permutation_solver}
\title{permutation_solver}
\usage{
permutation_solver(
  pattern,
  permutation_patterns,
  full_patterns = NULL,
  verbose = TRUE
)
}
\arguments{
\item{pattern}{A numeric vector where each element corresponds to the length of a run.}

\item{permutation_patterns}{A matrix of permutations as generated by \code{make_full_perm_set()}}

\item{full_patterns}{A list where each element is a numerical run length encoding corresponding
to the rows in the \code{permutation_patterns} matrix. If \code{NULL} (default), the run length
encodings will be calculated by the function.}
}
\value{
A matrix with the same number of columns as the \code{permutation_patterns}
argument, and the same number of rows as there are matches.
}
\description{
Given a row or column pattern and the set of all possible
permutations of the row/column's length, \code{permutation_solver} identifies
the set of permutations that satisfy the pattern
}
\examples{
full <- make_full_perm_set(10)
permutation_solver(c(2, 4, 1), full)
}
