% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nonogram.R
\name{plot.nonogram}
\alias{plot.nonogram}
\title{plot.nonogram}
\usage{
\method{plot}{nonogram}(nonogram, with_solution = TRUE)
}
\arguments{
\item{nonogram}{}

\item{with_solution}{}
}
\value{

}
\description{
Method for plotting solved and unsolved nonograms.
}
