% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.nonogram.R
\name{solve.nonogram}
\alias{solve.nonogram}
\title{solve.nonogram}
\usage{
\method{solve}{nonogram}(
  nonogram,
  row_permutations = NULL,
  column_permutations = NULL,
  verbose = TRUE,
  max_iter = 100
)
}
\arguments{
\item{nonogram}{}

\item{row_permutations}{}

\item{column_permutations}{}

\item{verbose}{}

\item{max_iter}{}
}
\value{

}
\description{
Solves nonograms blah blah blah
}
